[ningpop] : add READ function of session app and moduling initialize firebase
