+++ session/templates/session_create.html
@@ -1,0 +1,104 @@
+<!DOCTYPE html>
+<html lang="en">
+
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <link rel="stylesheet" href="../../base/reset.css">
+    <link rel="stylesheet" href="../../base/base.css">
+    <link rel="stylesheet" href="session_create.css">
+    <title>멋쟁이사자처럼 at 순천향대학교</title>
+</head>
+
+<body>
+    <header>
+        <div class="header-box">
+            <img src="../../image/mainLogo.png" alt="mainLogo.png">
+            <nav>
+                <ul>
+                    <li><a href="#">ABOUT</a></li>
+                    <li>
+                        <a href="#">NOTICE</a>
+                        <ul class="dropbox-nav">
+                            <li><a href="#">calendar</a></li>
+                            <li><a href="#">session</a></li>
+                        </ul>
+                    </li>
+                    <li>
+                        <a href="#">ACTIVITY</a>
+                        <ul class="dropbox-nav">
+                            <li><a href="#">curiculum</a></li>
+                            <li><a href="#">projects</a></li>
+                        </ul>
+                    </li>
+                    <li><a href="#">GALLARY</a></li>
+                    <li><a href="#">ASSIGNMENT</a></li>
+                    <li><a href="#">FAQ</a></li>
+                    <a class="login-box" href="#">LOGIN</a>
+                </ul>
+            </nav>
+        </div>
+    </header>
+
+    <main>
+
+        <!-- 서브 타이틀 필요 하신분은 그대로 쓰시고 아니면 지우시고 하시면 됩니다. -->
+        <div class='title-wrap'>
+            <span class='page-sub-title'>
+                SESSION
+            </span>
+        </div>
+
+        <!-- form -->
+
+        <div class="main-wrap">
+            <form class="form-box" action="#">
+                <div class="title-box">
+                    <p>제목</p>
+                    <input type="text" name="" placeholder="ex. html / css 를 이용한 네이버 실시간 검색어 클론코딩">
+                </div>
+                <div class="date-box">
+                    <p>세션 일시</p>
+                    <input type="date" name="">
+                </div>
+                <div class="link-box">
+                    <p>구글 문서 링크</p>
+                    <input type="text" name="" placeholder="ex. html / css 를 이용한 네이버 실시간 검색어 클론코딩">
+                </div>
+                <div class="explanation-box">
+                    <p>부가 설명</p>
+                    <textarea name="" cols="30" rows="10"></textarea>
+                </div>
+                <input type="submit" value="SUBMIT">
+            </form>
+        </div>
+
+        <!-- ------- -->
+
+
+    </main>
+
+    <footer>
+        <div class="footer-box">
+            <div class="contact-box">
+                <label>CONTACT US</label>
+                <!-- 인스타, 페북, 오픈카톡 연락처-->
+                <div class="icons">
+                    <a href="#"><i class="fas fa-comment-dots"></i></a>
+                    <a href="#"><i class="fab fa-instagram-square"></i></a>
+                    <a href="#"><i class="fab fa-facebook-square"></i></a>
+                </div>
+            </div>
+            <div class="copyright-text">
+                <span>Copyright ©</span> likelion at SCH 8th
+            </div>
+            <img src="../../image/slogan2.png" alt="slogan2.png">
+        </div>
+    </footer>
+
+    <script src="https://kit.fontawesome.com/54f85ae94e.js" crossorigin="anonymous"></script>
+    <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.2.1/jquery.min.js"></script>
+    <script src="script/jquery-1.12.3.js"></script>
+</body>
+
+</html>
