+++ session/templates/session_detail.html
@@ -1,0 +1,117 @@
+<!DOCTYPE html>
+<html lang="en">
+
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <link rel="stylesheet" href="../../base/reset.css">
+    <link rel="stylesheet" href="../../base/base.css">
+    <link rel="stylesheet" href="session_detail.css">
+    <title>멋쟁이사자처럼 at 순천향대학교</title>
+</head>
+
+<body>
+    <header>
+        <div class="header-box">
+            <img src="../../image/mainLogo.png" alt="mainLogo.png">
+            <nav>
+                <ul>
+                    <li><a href="#">ABOUT</a></li>
+                    <li>
+                        <a href="#">NOTICE</a>
+                        <ul class="dropbox-nav">
+                            <li><a href="#">calendar</a></li>
+                            <li><a href="#">session</a></li>
+                        </ul>
+                    </li>
+                    <li>
+                        <a href="#">ACTIVITY</a>
+                        <ul class="dropbox-nav">
+                            <li><a href="#">curiculum</a></li>
+                            <li><a href="#">projects</a></li>
+                        </ul>
+                    </li>
+                    <li><a href="#">GALLARY</a></li>
+                    <li><a href="#">ASSIGNMENT</a></li>
+                    <li><a href="#">FAQ</a></li>
+                    <a class="login-box" href="#">LOGIN</a>
+                </ul>
+            </nav>
+        </div>
+    </header>
+
+    <main>
+
+        <!-- 서브 타이틀 필요 하신분은 그대로 쓰시고 아니면 지우시고 하시면 됩니다. -->
+        <div class='title-wrap'>
+            <span class='page-sub-title'>
+                세션의 제목이 들어가는 공간
+            </span>
+        </div>
+
+        <!-- 과제 디테일을 보여주는 부분 -->
+
+        <div class="main-wrap">
+            <div class="detail-box">
+                <!-- detail 요소가 보여지는 내용 -->
+                <div class="update-delete-wrap">
+                    <a href="#">삭제</a>
+                    <a href="#">수정</a>
+                </div>
+
+                <div class="info-box">
+                    <div class="info-title">
+                        <span>INFO</span>
+                    </div>
+                    <span>SESSION DATE : <div class="text-bold">2021.03.02</div><br>
+                    </span>
+                    <span><a href="#">GOOGLE DOCUMENT LINK</a></span>
+                </div>
+
+                <div class="detail-box mg-top-60">
+                    <div class="detail-title">
+                        <span>DETAIL</span>
+                    </div>
+                    <!-- detail contents 가 들어가는 부분 -->
+                    <div class="detail-contents">
+                        세션과 관련한 자세한 내용들이 들어가는 공간 <br>
+                        ck 에디터로 작성된 아이들이 들어갈 공간 임니디.<br>
+                        어쩌구<br>
+                        저쩌구<br>
+                        씰룩<br>
+                        씰룩 ^~^<br>
+                    </div>
+                </div>
+
+                <!-- ------------------------------- -->
+
+            </div>
+        </div>
+
+
+    </main>
+
+    <footer>
+        <div class="footer-box">
+            <div class="contact-box">
+                <label>CONTACT US</label>
+                <!-- 인스타, 페북, 오픈카톡 연락처-->
+                <div class="icons">
+                    <a href="#"><i class="fas fa-comment-dots"></i></a>
+                    <a href="#"><i class="fab fa-instagram-square"></i></a>
+                    <a href="#"><i class="fab fa-facebook-square"></i></a>
+                </div>
+            </div>
+            <div class="copyright-text">
+                <span>Copyright ©</span> likelion at SCH 8th
+            </div>
+            <img src="../../image/slogan2.png" alt="slogan2.png">
+        </div>
+    </footer>
+
+    <script src="https://kit.fontawesome.com/54f85ae94e.js" crossorigin="anonymous"></script>
+    <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.2.1/jquery.min.js"></script>
+    <script src="script/jquery-1.12.3.js"></script>
+</body>
+
+</html>
