+++ static/css/session_create.css
@@ -1,0 +1,107 @@
+@font-face {
+    font-family: 'IBMPlexSansKR-Text';
+    src: url('https://cdn.jsdelivr.net/gh/projectnoonnu/noonfonts_20-07@1.0/IBMPlexSansKR-Text.woff') format('woff');
+    font-weight: normal;
+    font-style: normal;
+}
+
+@font-face {
+    font-family: 'IBMPlexSansKR-Light';
+    src: url('https://cdn.jsdelivr.net/gh/projectnoonnu/noonfonts_20-07@1.0/IBMPlexSansKR-Light.woff') format('woff');
+    font-weight: normal;
+    font-style: normal;
+}
+
+@font-face {
+    font-family: 'IBMPlexSansKR-Bold';
+    src: url('https://cdn.jsdelivr.net/gh/projectnoonnu/noonfonts_20-07@1.0/IBMPlexSansKR-Bold.woff') format('woff');
+    font-weight: normal;
+    font-style: normal;
+}
+
+:root {
+    --main-width: 1100px;
+    --SCH-skyblue: #1BBCF0;
+}
+
+* {
+    text-decoration: none;
+}
+
+/********************************************************/
+
+
+/* form box */
+
+.main-wrap {
+    width: 100%;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+}
+
+.form-box {
+    width: 900px;
+    height: auto;
+    margin-top: 30px;
+}
+
+.form-box p {
+    font-size: 20px;
+    font-family: "IBMPlexSansKR-Text";
+}
+
+.title-box input,
+.link-box input,
+.date-box input {
+    width: 100%;
+    height: 35px;
+    padding-left: 20px;
+    margin-top: 15px;
+    outline: none;
+    border: 1px solid rgb(228, 228, 228);
+    border-radius: 10px;
+}
+
+input::placeholder {
+    color: #E6E6E6;
+}
+
+.date-box,
+.explanation-box,
+.link-box {
+    margin-top: 50px;
+}
+
+.date-box input {
+    width: 200px;
+}
+
+.explanation-box textarea {
+    font-family: "IBMPlexSansKR-Lightt";
+    width: 100%;
+    height: 200px;
+    overflow: auto;
+    margin-top: 15px;
+    border: 1px solid rgb(228, 228, 228);
+    border-radius: 10px;
+    outline: none;
+    padding: 20px;
+}
+
+.form-box>input[type=submit] {
+    width: 100%;
+    height: 50px;
+    margin-top: 50px;
+    background-color: white;
+    border: 2px solid var(--SCH-skyblue);
+    color: var(--SCH-skyblue);
+    border-radius: 15px;
+    transition: .4s;
+    margin-bottom: 60px;
+}
+
+.form-box>input[type=submit]:hover {
+    background-color: var(--SCH-skyblue);
+    color: white;
+}
