+++ static/css/session_detail.css
@@ -1,0 +1,98 @@
+@font-face {
+    font-family: 'IBMPlexSansKR-Text';
+    src: url('https://cdn.jsdelivr.net/gh/projectnoonnu/noonfonts_20-07@1.0/IBMPlexSansKR-Text.woff') format('woff');
+    font-weight: normal;
+    font-style: normal;
+}
+
+@font-face {
+    font-family: 'IBMPlexSansKR-Light';
+    src: url('https://cdn.jsdelivr.net/gh/projectnoonnu/noonfonts_20-07@1.0/IBMPlexSansKR-Light.woff') format('woff');
+    font-weight: normal;
+    font-style: normal;
+}
+
+@font-face {
+    font-family: 'IBMPlexSansKR-Bold';
+    src: url('https://cdn.jsdelivr.net/gh/projectnoonnu/noonfonts_20-07@1.0/IBMPlexSansKR-Bold.woff') format('woff');
+    font-weight: normal;
+    font-style: normal;
+}
+
+:root {
+    --main-width: 1100px;
+    --SCH-skyblue: #1BBCF0;
+}
+
+* {
+    text-decoration: none;
+}
+
+/********************************************************/
+
+.main-wrap {
+    width: 100%;
+    height: auto;
+    display: flex;
+    justify-content: center;
+}
+
+.detail-box {
+    width: var(--main-width);
+    height: auto;
+    margin-top: 50px;
+}
+
+.update-delete-wrap {
+    display: flex;
+    flex-direction: row-reverse;
+}
+
+.update-delete-wrap a {
+    padding: 0px 5px;
+    color: var(--SCH-skyblue);
+}
+
+.info-title,
+.link-title,
+.detail-title {
+    width: 100%;
+    padding-bottom: 1px;
+    margin-top: 40px;
+    border-bottom: 1px solid rgb(228, 228, 228);
+}
+
+.info-title span,
+.link-title span,
+.detail-title span {
+    border-bottom: 2px solid lightgray;
+    font-family: 'IBMPlexSansKR-Text';
+    font-size: 24px;
+}
+
+.info-box>span,
+.link-box>a {
+    font-size: 20px;
+    width: 50px;
+    line-height: 50px;
+}
+
+.link-box>a {
+    text-decoration: underline;
+}
+
+.detail-contents {
+    margin: 25px 0px 60px 0px;
+    font-size: 20px;
+}
+
+
+/* **************************** */
+.text-bold {
+    display: inline;
+    font-family: 'IBMPlexSansKR-Text';
+}
+
+.mg-top-60 {
+    margin-top: 60px;
+}
